requirements.txt
pandas
numpy
yfinance
confluent_kafka
streamlit
plotly.express